feat: StockInfo 컴포넌트와 로직 코드 분리 (#36)
